- raise ConfigException when the checkout mismatches with the importer config   - check that the source dir is a CVS checkout before doing anything else
git-svn-id: svn+ssh://rubyforge.org/var/svn/autobuild/trunk/autobuild@213 b7efc242-7007-0410-a639-993fcd5ee89a

+++ lib/autobuild/import/cvs.rb
@@ -24,11 +24,15 @@ module Autobuild
 
         def update(package)
             Dir.chdir(package.srcdir) do
+		if !File.exists?("#{package.srcdir}/CVS/Root")
+		    raise ConfigException, "#{package.srcdir} does not look like a CVS checkout"
+		end
+
 		root = File.open("#{package.srcdir}/CVS/Root") { |io| io.read }.chomp
 		mod  = File.open("#{package.srcdir}/CVS/Repository") { |io| io.read }.chomp
 
 		if root != @root || mod != @module
-		    raise ArgumentError, "checkout in #{package.srcdir} is from #{root}:#{mod}, was expecting #{@root}:#{@module}"
+		    raise ConfigException, "checkout in #{package.srcdir} is from #{root}:#{mod}, was expecting #{@root}:#{@module}"
 		end
                 Subprocess.run(package.name, :import, @program, 'up', *@options_up)
 	    end
